requirements.txt
streamlit==1.33.0

langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38

langchain-groq==0.1.9

SQLAlchemy==2.0.29

python-dotenv==1.0.1


pydantic==2.7.4